+++ src/sipnet/limitations.c
@@ -9,6 +9,61 @@
 #include "nitrogen.h"
 #include "state.h"
 
+// See limitations.h
+void checkLeafOnLimitation(double *leafOnFlux) {
+  // Leaf on events are limited by:
+  // * leafGrowth parameter (input leafOn value)
+  // * available carbon
+  // * available nitrogen
+  double leafOnCDemand = *leafOnFlux * climate->length;
+
+  if (leafOnCDemand < TINY) {
+    // Nothing to check
+    return;
+  }
+
+  // First up, carbon. We do not draw from the C storage pool for this.
+  double availableC =
+      (envi.plantWoodC + envi.coarseRootC) * params.leafOnReallocFrac;
+  double cLimiter = availableC / leafOnCDemand;
+
+  double leafOnNDemand = 0.0;
+  double nLimiter = 1.0;
+  double availableN = 0.0;
+  // Next, nitrogen; leaf-on only draws from the plantStorageN pool
+  if (ctx.nitrogenCycle) {
+    // Needed N for this transfer is (what leaves need) - (what wood provides)
+    // Reminder: both wood and coarseRoot use params.woodCN, so no need to
+    // treat those C demands separately
+    leafOnNDemand = calcLeafOnNFromC(leafOnCDemand);
+    availableN = envi.plantStorageN;
+    if (leafOnNDemand > TINY) {
+      nLimiter = availableN / leafOnNDemand;
+    }
+  }
+  double limitation = fmin(cLimiter, nLimiter);
+  limitation = fmax(fmin(limitation, 1.0), 0.0);
+
+  if (limitation < 1) {
+    *leafOnFlux *= limitation;
+    if (ctx.nitrogenCycle) {
+      logInfo("Leaf on creation %.4f C / %.4f N exceeds available C %.4f / N "
+              "%.4f (C ratio: %.4f, N ratio: %.4f), "
+              "reducing leaf-on growth by %.2f%% on year %d day %d time %.3f\n",
+              leafOnCDemand, leafOnNDemand, availableC, availableN, cLimiter,
+              nLimiter, (1 - limitation) * 100, climate->year, climate->day,
+              climate->time);
+    } else {
+      logInfo("Leaf on creation %.4f exceeds available C %.4f "
+              "(C ratio: %.4f, N ratio: %.4f), "
+              "reducing leaf-on growth by %.2f%% on year %d day %d time %.3f\n",
+              leafOnCDemand, availableC, cLimiter, nLimiter,
+              (1 - limitation) * 100, climate->year, climate->day,
+              climate->time);
+    }
+  }
+}
+
 /**
  * Check for nitrogen limitation, and reduce growth if needed
  */
@@ -17,9 +72,7 @@ static void checkNitrogenLimitation(void) {
   double maxDemandFlux = calcPlantNDemand();
   double maxDemand = maxDemandFlux * climate->length;
 
-  double nonUptakeFluxes = calcMinNNonUptakeFluxes();
-  double availableMinN =
-      fmax(0, envi.minN + (nonUptakeFluxes * climate->length));
+  double availableMinN = calcPlantAvailableN();
 
   double nFixationFrac = calcNFixationFrac();
   double maxUptake = maxDemand * (1 - nFixationFrac);
@@ -32,9 +85,8 @@ static void checkNitrogenLimitation(void) {
             maxUptake, availableMinN, (1 - reduction) * 100, climate->year,
             climate->day, climate->time);
 
-    // Reduce all drains on soil N
-    fluxes.eventLeafOnCreation *= reduction;
-    fluxes.leafOnCreation *= reduction;
+    // Reduce all drains on soil N (all fluxes used in calcPlantNDemand, plus
+    // fixation and uptake)
     fluxes.woodCreation *= reduction;
     fluxes.leafCreation *= reduction;
     fluxes.fineRootCreation *= reduction;
@@ -46,8 +98,7 @@ static void checkNitrogenLimitation(void) {
 
 // See limitations.h
 void checkLimitations(void) {
-  // EXPECTED: calcLeafOnLimitation()
-
+  // Our only post-flux limitation to check
   if (ctx.nitrogenCycle) {
     checkNitrogenLimitation();
   }